More Ids

+++ Octokit/Models/Response/RepositoryInvitation.cs
@@ -22,7 +22,7 @@ namespace Octokit
     {
         public RepositoryInvitation() { }
 
-        public RepositoryInvitation(int id, string nodeId, Repository repository, User invitee, User inviter, InvitationPermissionType permissions, DateTimeOffset createdAt, string url, string htmlUrl)
+        public RepositoryInvitation(long id, string nodeId, Repository repository, User invitee, User inviter, InvitationPermissionType permissions, DateTimeOffset createdAt, string url, string htmlUrl)
         {
             Id = id;
             NodeId = nodeId;
@@ -35,7 +35,7 @@ namespace Octokit
             HtmlUrl = htmlUrl;
         }
 
-        public int Id { get; protected set; }
+        public long Id { get; protected set; }
 
         /// <summary>
         /// GraphQL Node Id
